added the ability to confirm delete before remove task, along with fixing issue of delete button causing red toggle

+++ server/public/scripts/client.js
@@ -6,7 +6,14 @@ $(document).ready(function () {
 
 function setupClickListeners() {
   $('#view-todo').on('click', '.btn-delete', deleteTodo);
-  $('#view-todo').on('click', '.todo-list', checkTodo);
+  $('#view-todo').on('click', '.todo-list', function (e) {
+    let id = $(this).closest('section').data('id');
+    let taskStatus = $(this).closest('section').data('status');
+    let target = $(e.target);
+    if (target.is('p')) {
+      checkTodo(id, taskStatus);
+    }
+  });
   $('#submit-btn').on('click', saveTodo);
 }
 
@@ -64,8 +71,8 @@ function renderTodo(todos) {
     $('#view-todo').append(`
       <section class='todo-task' data-id='${todo.id}' data-status='${todo.status}'>
       <div class='todo-list' id='${todo.id}'>
-        <p>${todo.task}</p>
-        <p>Status: ${todo.status}</p>
+        <p class='todo-change'>${todo.task}</p>
+        <p class='todo-change'>Status: ${todo.status}</p>
         <p><button class='btn-delete'><img src='${deleteImg}'></button></p>
         </div>
         </section>
@@ -79,26 +86,46 @@ function renderTodo(todos) {
 
 function deleteTodo() {
   let todoId = $(this).closest('section').data('id');
-  $.ajax({
-    method: 'DELETE',
-    url: `/todo/${todoId}`,
-  })
-    .then((res) => {
-      getTodo();
-    })
-    .catch((error) => {
-      console.error('Error in delete', error);
-      alert('Could not delete task, please try again.');
-    });
+  Swal.fire({
+    title: 'Are you sure?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    width: 400,
+    position: 'top',
+    background: '#1d3557',
+    showCancelButton: true,
+    confirmButtonText: 'Yes, delete it!',
+  }).then((res) => {
+    if (res.isConfirmed) {
+      Swal.fire({
+        title: 'Deleted!',
+        confirmButtonText: 'Your todo task is no more!',
+        position: 'top',
+        background: '#1d3557',
+        width: 400,
+        icon: 'success',
+      });
+      $.ajax({
+        method: 'DELETE',
+        url: `/todo/${todoId}`,
+      })
+        .then((res) => {
+          getTodo();
+        })
+        .catch((error) => {
+          console.error('Error in delete', error);
+          alert('Could not delete task, please try again.');
+        });
+    }
+  });
 }
 
-function checkTodo() {
-  let id = $(this).closest('section').data('id');
-  let taskStatus = $(this).closest('section').data('status');
+function checkTodo(id, status) {
+  console.log(id, status);
   $.ajax({
     method: 'PUT',
     url: `/todo/${id}`,
-    data: { taskStatus },
+    data: { status },
   })
     .then((res) => {
       getTodo();
